feat: refactor policy-management in it's own application service
