+++ apps/main/src/identity/organizations/application/services/organizations.service.ts
@@ -19,7 +19,7 @@ import { InvitationsRepository } from "../../infrastructure/adapters/invitations
 import { MembersRepository } from "../../infrastructure/adapters/members.repository";
 import { OrganizationsRepository } from "../../infrastructure/adapters/organizations.repository";
 import { InstanceSettingsService } from "../../../../instance-settings/application/services/instance-settings.service";
-import { PolicyService } from "../../../../policy/infrastructure/policy.service";
+import { PolicyManagementService } from "../../../../policy/application/services/policy-management.service";
 import { InvitationPopulateDecorator } from "../invitation-populate-decorator";
 import type { InvitationResponseDto, InvitationStatusDtoType } from "@open-dpp/dto";
 
@@ -33,7 +33,7 @@ export class OrganizationsService {
     private readonly usersRepository: UsersRepository,
     private readonly invitationsRepository: InvitationsRepository,
     private readonly instanceSettingsService: InstanceSettingsService,
-    private readonly policyService: PolicyService,
+    private readonly policyManagementService: PolicyManagementService,
   ) {}
 
   async createOrganization(
@@ -57,7 +57,7 @@ export class OrganizationsService {
       throw new BadRequestException();
     }
     try {
-      await this.policyService.ensureDefaultPolicies(createdOrganization.id);
+      await this.policyManagementService.ensureDefaultPolicies(createdOrganization.id);
     } catch (error) {
       this.logger.error(
         `Failed to seed default policies for organization ${createdOrganization.id}; ` +
